+++ FiapTechChallenge.ContactBook.Domain.Core/Entities/DDD.cs
@@ -5,7 +5,7 @@ namespace FiapTechChallenge.ContactBook.Domain.Core.Entities
     public class DDD : BaseEntity
     {
         public string Code { get; set; }
-        public int RegionId { get; set; }
+        public Int64 RegionId { get; set; }
 
         public virtual Region RegionNavigation { get; set; }
